[NOT WORKING] Implement Presidential Poll Spider
This presidential poll spider implementation extends the CrawlSpider
class.  Scrapy recommends this for most crawling spiders, but it doesn't
seem to work in this case.  The main issue is that none of the links
matched by the specified regex are followed, nor are any of the callback
functions called.  It could be due to a semantic error in the pres2012
spider or an issue with Scrapy itself.