Add State Field, Fix Date Fields for PresPollItems
Previously the date fields for PresPollItems both contained the poll's
full date range (ex 8/17 - 8/20).  Now, the start field contains the
poll's start date and the end field contains the poll's end date.  Note
that the end date is the one used to determine the 'day' a poll is for.

Add a state field to each item.  This is useful in post-processing of
the data items, to seperate them into the right files.