Fix Already Seen Prevention in CsvExportPipeline
Handle cases when data/dict.json doesn't exist (ie when the spider is
run for the first time).

Correctly write json output so it can be read.  The problem was that
a file written with str(data) couldn't be read in with json.load- so I
changed str(data) to json.dumps(data).  This caused another error to
emerge- json.dumps(data) wouldn't dump a dict with hash:pollItem as the
key:value pair.  Since the actual pollItem doesn't need to be stored,
just the fact that it was seen in a previous crawl, hash:True can be
used for the key:value pair without a loss of functionality.