Correctly Match RCP Poll Headers to Item Attributes
RealClearPolitics.com state-level poll tables do not have consistent
ordering: the candidates are listed from left to right based on who
ranks highest in the RCP average.

This update removes static field selection, which presumes consistent
ordering (ie the democratic candidate's points will always be in fifth
column of the table).  Instead, it creates a dynamic ordering in the
_getDataPosition function by matching table header position to an item's
attribute.