Update Documentation, Fix Bugs, Handle States with no Polls
Add documentation for pres2012 spider's _parsePollDates method.  Fix a
bug in that method, when only one date is present.

Prevent pres2012 spider from visiting states with no available polling
data.  This is done in the processLinks method by checking links against
a default link (which indicates no polls are available for that state).