Refactor Header-To-Attribute Functionality
Use a utility module (in polldata/utils/) to handle the mapping between
a state page's Poll Data table headers to the pollitem's attributes.
Note that the mapping is not exact, and some mappings consolidate
attributes (ie the 'Date' => 'date' mapping consolidates pollitem's
start and end attributes into date).

The use of a utility class makes it easy to change the mappings accross
all spiders, if needed.