Change Data Source to Find All States with Presidential Polls
Previously, the pres2012 spider started at
http://www.realclearpolitics.com/epolls/latest_polls/president/
and crawled the state pages that appeared in that latest-polls list.
Unfortunately the latest polls page occasionally lags behind individual
state pages (they are updated first).

Instead of crawling the latest polls page and only visiting states that
are listed on it, the pres2012 spider should visit every state page with
presidential polls.  This ensures that all of the latest polls are
found.  The challenge in this approach is finding the links: a visit to
the latest polls page shows a select box that lists all of the state
pages with presidential polls- however, that select box is created from
javascript when the page loads.

There are two options to extract the urls from the select box:
 1. extract the urls directly from the javascript (which contains
    obfuscated html), on
    http://www.realclearpolitics.com/epolls/2012/widget/search_by_race.js
 2. load the latest polls page and execute the javascript, then attempt
    to extract the links from the generated select box

Option 2. results in the cleanest spider code, but involved another
program to load the html and execute the js.  This increases the
complexity of the system enough that option 1. is more viable and, since
the system will be less complex, easier to maintain for future students
working with this codebase.

So, what acutally happened?
  - Create a new link extractor to handle the obfuscated html and parse
    all links.  To do this I copied the existing RegexLinkExtractor code
    and modified the regex and cleaning functions to handle the expected
    input.
  - Override the CrawlSpider _request_to_follow function so that
    non-html links are processed by the Rules.