Create a Function to Handle the Header-to-Attribute Lookup
The previous lookup method (by using the utils.headerToAttribute
variable directly) does not handle edge cases very well.

Create a function that mimics this functionality but allows for more
flexibility in the lookup process.  For example, the function coud do
some preprocessing of the 'header' input to generalize the lookup for
'dem' and 'rep' attributes by removing the candidate name headers.

Actually, that example sounds good.  This commit also includes that
functionality.