Spiders Should Work When New Polls Come Out
This fixes a bug that occurred when a spider was ran a second time after
new polls had come out.  The list of polls would be appended to
data/spider_dict.json rather than overwriting it.  This would create
malformed json and crash the spider when it was ran afterwards.